+++ app/containers/ActionSheet/styles.ts
@@ -63,5 +63,9 @@ export default StyleSheet.create({
 	},
 	rightContainer: {
 		paddingLeft: 12
+	},
+	footerButtonsContainer: {
+		flexDirection: 'row',
+		paddingTop: 16
 	}
 });
